+++ src/lib.rs
@@ -206,5 +206,5 @@ impl Mobi {
 /// Trait allowing generic reading of header fields
 pub(crate) trait HeaderField {
     /// Returns a position in the text where this field can be read
-    fn position(self) -> u16;
+    fn position(self) -> u64;
 }
